pyproject.toml
[project]
name = "status-page"
version = "0.1.0"
description = "Simple infrastructure status page"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "dependency-injector>=4.48.3",
    "fastapi>=0.128.0",
    "htmlmin>=0.1.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "scalar-fastapi>=1.6.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "bandit>=1.9.2",
    "pre-commit>=4.5.1",
    "pyright>=1.1.407",
    "ruff>=0.14.10",
]

[tool.ruff]
src = ["src"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
]
line-length = 79
fix = true
unsafe-fixes = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["INP001", "ARG001"]

[tool.pyright]
include = ["src"]
exclude = [
    "**/__pycache__",
    ".venv",
]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingTypeStubs = false
